+++ CHANGELOG.md
@@ -14,6 +14,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - Settings from Slew and Center button
+## [App4.2.1]
+### Fixed
+- Fixed shutdown and restart
 
 ## [App4.2.0] [livestack0.6.2][bahtifocus1.0.0] - 2025-12-04
 ### Important information